+++ Integral/IntRuleHexahedron.h
@@ -39,7 +39,7 @@ class IntRuleHexahedron : public IntRule
     virtual void SetOrder(int order) override;
     
     // Integration rule 3D (hexahedron elements) method obtained from Numerical Recipes
-    static void gaulegHexa(const double x1, const double x2, VecDouble &x, VecDouble &w);
+    static void gaulegHexa(const REAL x1, const REAL x2, VecDouble &x, VecDouble &w);
 
 };
