+++ Matrix/SparseEigen.h
@@ -2,10 +2,11 @@
 #define SPARSE_EIGEN_H
 
 #include <Eigen/SparseCore>
+#include "DenseEigen.h"
 
 using namespace Eigen;
 
 /// @brief Sparse matrix of doubles
-typedef Eigen::SparseMatrix<double> SparseMat;
+typedef Eigen::SparseMatrix<REAL> SparseMat;
 
 #endif
